elf2tbf: convert to elf2tab
elf2tbf now creates TAB files so we can get rid of the python script. It
also means we don't have to run elf2tbf multiple times.
