+++ userland/Configuration.mk
@@ -32,11 +32,11 @@ PACKAGE_NAME ?= $(shell basename "$(shell pwd)")
 # Tock supported architectures
 TOCK_ARCHS ?= cortex-m0 cortex-m3 cortex-m4
 
-# This could be replaced with an installed version of `elf2tbf`
-ELF2TBF ?= cargo run --manifest-path $(TOCK_USERLAND_BASE_DIR)/tools/elf2tbf/Cargo.toml --
-ELF2TBF_ARGS += -n $(PACKAGE_NAME)
-ELF2TBF_ARGS += --stack $(STACK_SIZE) --app-heap $(APP_HEAP_SIZE) --kernel-heap $(KERNEL_HEAP_SIZE)
-ELF2TBF_ARGS += --crt0-header
+# This could be replaced with an installed version of `elf2tab`
+ELF2TAB ?= cargo run --manifest-path $(TOCK_USERLAND_BASE_DIR)/tools/elf2tab/Cargo.toml --
+ELF2TAB_ARGS += -n $(PACKAGE_NAME)
+ELF2TAB_ARGS += --stack $(STACK_SIZE) --app-heap $(APP_HEAP_SIZE) --kernel-heap $(KERNEL_HEAP_SIZE)
+ELF2TAB_ARGS += --crt0-header
 
 # Flags for building app Assembly, C, C++ files
 # n.b. make convention is that CPPFLAGS are shared for C and C++ sources
