+++ userland/examples/blink/main.c
@@ -5,8 +5,6 @@ int main(void) {
   // Ask the kernel how many LEDs are on this board.
   int num_leds = led_count();
 
-  num_leds = buf[0];
-
   // Blink the LEDs in a binary count pattern and scale
   // to the number of LEDs on the board.
   for (int count = 0; ; count++) {
