+++ userland/tools/elf2tab/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "elf2tab"
+version = "0.3.0"
+description = "Compiles from ELF to TAB (a Tock Application Bundle using the Tock Binary Format)"
+authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
+
+[dependencies]
+getopts = "0.2"
+chrono = "0.4.2"
+tar = "0.4.14"
+elf = { git = "https://github.com/cole14/rust-elf" }
+
