+++ userland/tools/elf2tab/README.md
@@ -1,0 +1,3 @@
+# elf2tab
+
+A compiler from ELF to TAB, using the [Tock Binary Format](../../../doc/Compilation.md#tock-binary-format).
