--- userland/tools/elf2tbf/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "elf2tbf"
-version = "0.2.0"
-description = "Compiles from ELF to TBF (the Tock Binary Format)"
-authors = ["Tock Project Developers <tock-dev@googlegroups.com>"]
-
-[dependencies]
-getopts = "0.2"
-elf = { git = "https://github.com/cole14/rust-elf" }
-
