--- userland/tools/elf2tbf/README.md
@@ -1,3 +1,0 @@
-# elf2tbf
-
-A compiler from ELF to TBF ([Tock Binary Format](../../../doc/Compilation.md#tock-binary-format)).
