+++ userland/userland_generic.ld
@@ -75,7 +75,7 @@ SECTIONS {
      * .ARM.exidx is sorted, so has to go in its own output section.
      *
      * __NOTE__: It's at the end because we currently don't actually serialize
-     * it to the binary in elf2tbf. If it was before the RAM sections, it would
+     * it to the binary in elf2tab. If it was before the RAM sections, it would
      * through off our calculations of the header.
      */
     PROVIDE_HIDDEN (__exidx_start = .);
